requirements.txt
pypdf==6.0.0
openai==1.63.0
pdf2image==1.17.0
python-dotenv==1.0.1